+++ application/services/RelationValidator/IRelationValidator.go
@@ -1,6 +1,6 @@
 package RelationValidator
 
 type IRelationValidator interface {
-	GenerateToken(userID int, relationType int, idx int) (string, error)
-	GetTokenInfo(token string) (userID int, relationType int, err error)
+	GenerateToken(userID int64, relationType int, idx int) (string, error)
+	GetTokenInfo(token string) (userID int64, relationType int, err error)
 }
