+++ application/services/relationship.go
@@ -10,20 +10,20 @@ import (
 
 type RelationShipService interface {
 	//UserHasAssociation will return a boolean value if the user has an association with other users
-	UserHasAssociation(userID int, type_ int) bool
+	UserHasAssociation(userID int64, type_ int) bool
 
 	//CreateRelationshipConnection will return a string value which it should be an url to the user's relationship connection
-	CreateRelationshipConnection(userID int, type_ int) string
+	CreateRelationshipConnection(userID int64, type_ int) string
 
 	// DisassociateUser will remove the user's association with other users
-	DisassociateUser(userID int, associateID int) bool
+	DisassociateUser(userID int64, associateID int64) bool
 
 	// BindingTwoUser will bind two users with the relationship
-	BindingTwoUser(link string, receiverID int) (err error)
+	BindingTwoUser(link string, receiverID int64) (err error)
 
-	GetFriendList(userID int) []models.Relationship
+	GetFriendList(userID int64) []models.Relationship
 
-	GetHistoryList(userID int) []models.RelationShipHistory
+	GetHistoryList(userID int64) []models.RelationShipHistory
 }
 
 type relationshipServiceImpl struct {
@@ -32,13 +32,13 @@ type relationshipServiceImpl struct {
 	validator RelationValidator.IRelationValidator
 }
 
-func (r *relationshipServiceImpl) GetFriendList(userID int) []models.Relationship {
+func (r *relationshipServiceImpl) GetFriendList(userID int64) []models.Relationship {
 	lists := r.repo.GetRelationList(userID)
 
 	return lists
 }
 
-func (r *relationshipServiceImpl) BindingTwoUser(link string, receiverID int) error {
+func (r *relationshipServiceImpl) BindingTwoUser(link string, receiverID int64) error {
 
 	senderID, relationType, err := r.validator.GetTokenInfo(link)
 	if err != nil {
@@ -68,7 +68,7 @@ func (r *relationshipServiceImpl) BindingTwoUser(link string, receiverID int) er
 // UserHasAssociation mean to check users has been associate relationship with others
 // if user want's to bind to other friend, just don't need to care
 // but if user wants to bind Couple relationship , this must be limited to 1
-func (r *relationshipServiceImpl) UserHasAssociation(userID int, type_ int) bool {
+func (r *relationshipServiceImpl) UserHasAssociation(userID int64, type_ int) bool {
 	count := r.UserAssociationCount(userID, type_)
 
 	if type_ == models.Couple {
@@ -78,13 +78,13 @@ func (r *relationshipServiceImpl) UserHasAssociation(userID int, type_ int) bool
 	return count < 10
 }
 
-func (r *relationshipServiceImpl) UserAssociationCount(userID int, type_ int) int {
+func (r *relationshipServiceImpl) UserAssociationCount(userID int64, type_ int) int {
 	return r.repo.GetRelationCount(userID, type_)
 }
 
 // CreateRelationshipConnection a method for create bind link for users, if this user can add more friend, method will return a linkpath
 // otherwise return empty string
-func (r *relationshipServiceImpl) CreateRelationshipConnection(userID int, type_ int) string {
+func (r *relationshipServiceImpl) CreateRelationshipConnection(userID int64, type_ int) string {
 	if r.UserHasAssociation(userID, type_) {
 		// todo:in this place, service will do twice sql check,this is not necessary need to be improved
 		count := r.UserAssociationCount(userID, type_)
@@ -100,7 +100,7 @@ func (r *relationshipServiceImpl) CreateRelationshipConnection(userID int, type_
 }
 
 // DisassociateUser this method will unbind the user's relationship with other users
-func (r *relationshipServiceImpl) DisassociateUser(userID int, associateID int) bool {
+func (r *relationshipServiceImpl) DisassociateUser(userID int64, associateID int64) bool {
 	if err := r.repo.UnBindTwoUser(userID, associateID); err != nil {
 		log.Error("DisassociateUser failed! error: ", err, " userID: ", userID, " associateID: ", associateID)
 		return false
@@ -108,7 +108,7 @@ func (r *relationshipServiceImpl) DisassociateUser(userID int, associateID int) 
 	return true
 }
 
-func (r *relationshipServiceImpl) GetHistoryList(userID int) []models.RelationShipHistory {
+func (r *relationshipServiceImpl) GetHistoryList(userID int64) []models.RelationShipHistory {
 	return r.repo.GetHistoryList(userID)
 }
